Week3-GCP/requirements.txt
# Function dependencies, for example:
# package>=version
pandas==1.4.2
google-cloud-storage==2.5.0
requests==2.27.1
